refactor(api): update ExtensionApi documentation and improve type definitions for clipboard history items
