+++ src/services/applicationsService.ts
@@ -1,34 +1,43 @@
 import { openPath } from "@tauri-apps/plugin-opener";
 import { invoke } from "@tauri-apps/api/core";
-import { info } from "@tauri-apps/plugin-log";
+import { LogService } from "./logService";
 import type { AppResult } from "../types";
 
+/**
+ * Service for managing and interacting with system applications
+ */
 class ApplicationsService {
-  private appCache: Map<string, string[]> = new Map();
-  private lastUpdate: number = 0;
-  private readonly CACHE_DURATION = 5000;
+  private appCache = new Map<string, string[]>();
+  private lastUpdate = 0;
+  private readonly CACHE_DURATION = 5000; // 5 seconds
 
+  /**
+   * Extract app name from path
+   */
   private getAppName(path: string): string {
     const parts = path.split("/");
     return parts[parts.length - 1].replace(".app", "");
   }
 
+  /**
+   * Refresh the application cache from the system
+   */
   async refreshCache(): Promise<void> {
     try {
       const paths: string[] = await invoke("list_applications");
       this.appCache.clear();
 
-      paths.forEach((path) => {
+      for (const path of paths) {
         const name = this.getAppName(path);
         if (!this.appCache.has(name)) {
           this.appCache.set(name, []);
         }
         this.appCache.get(name)!.push(path);
-      });
+      }
 
       this.lastUpdate = Date.now();
     } catch (error) {
-      info(`Failed to refresh application cache: ${String(error)}`);
+      LogService.error(`Failed to refresh application cache: ${error}`);
     }
   }
 
@@ -36,18 +45,20 @@ class ApplicationsService {
    * Ensures the application cache is loaded and up-to-date
    */
   async ensureCacheLoaded(): Promise<void> {
-    if (
+    const isCacheStale =
       this.appCache.size === 0 ||
-      Date.now() - this.lastUpdate > this.CACHE_DURATION
-    ) {
+      Date.now() - this.lastUpdate > this.CACHE_DURATION;
+
+    if (isCacheStale) {
       await this.refreshCache();
     }
   }
 
+  /**
+   * Search for applications matching the query
+   */
   async search(query: string): Promise<AppResult[]> {
-    if (Date.now() - this.lastUpdate > this.CACHE_DURATION) {
-      await this.refreshCache();
-    }
+    await this.ensureCacheLoaded();
 
     const results: AppResult[] = [];
     const searchTerm = query.toLowerCase();
@@ -57,6 +68,7 @@ class ApplicationsService {
         // Prefer /System/Applications path if available
         const bestPath =
           paths.find((p) => p.startsWith("/System/Applications")) || paths[0];
+
         results.push({
           name,
           path: bestPath,
@@ -68,9 +80,12 @@ class ApplicationsService {
     return results;
   }
 
+  /**
+   * Open an application
+   */
   async open(app: AppResult): Promise<void> {
     try {
-      invoke("hide");
+      invoke("hide"); // Hide the launcher window
       await openPath(app.path);
     } catch (error) {
       // If the first attempt fails, try the app name with .app extension
@@ -79,14 +94,13 @@ class ApplicationsService {
         invoke("hide");
         await openPath(altPath);
       } catch (retryError) {
-        info(`Failed to open ${app.name}: ${error}`);
+        LogService.error(`Failed to open ${app.name}: ${error}`);
       }
     }
   }
 
   /**
    * Gets all applications without filtering
-   * @returns All applications in the cache
    */
   public async getAllApplications(): Promise<AppResult[]> {
     await this.ensureCacheLoaded();
