+++ src/services/extensionDiscovery.ts
@@ -1,24 +1,24 @@
+import { LogService } from "./logService";
+
+// Use Vite's import.meta.glob to discover extensions
 const extensionContext = import.meta.glob("../extensions/*/manifest.json");
 
+/**
+ * Discover all available extensions in the project
+ */
 export async function discoverExtensions(): Promise<string[]> {
   try {
-    // LogService.error("Starting extension discovery process...");
-
-    // Get all extension paths from Vite's import.meta.glob
-    const extensionPaths = Object.keys(extensionContext);
-
     // Extract extension IDs from paths
-    const extensionIds = extensionPaths
+    const extensionIds = Object.keys(extensionContext)
       .map((path) => {
         const matches = path.match(/\/extensions\/(.+)\/manifest\.json/);
         return matches ? matches[1] : null;
       })
       .filter((id): id is string => id !== null);
 
-    // LogService.info(`Discovered ${extensionIds.length} extensions:`);
     return extensionIds;
-  } catch (err) {
-    LogService.error("No extensions found or error during discovery");
+  } catch (error) {
+    LogService.error(`Error during extension discovery: ${error}`);
     return [];
   }
 }
