+++ src/stores/clipboardHistoryStore.ts
@@ -3,12 +3,12 @@ import type { ClipboardHistoryItem } from "../types/clipboardHistoryItem";
 import { writable, type Writable } from "svelte/store";
 import { LogService } from "../services/logService";
 
-// Define store path
+// Constants
 const STORE_PATH = "clipboard_history.json";
-const MAX_HISTORY_AGE_MS = 90 * 24 * 60 * 60 * 1000; // 3 months in milliseconds
-const MAX_ITEMS = 1000; // Maximum number of items to store
+const MAX_HISTORY_AGE_MS = 90 * 24 * 60 * 60 * 1000; // 3 months
+const MAX_ITEMS = 1000;
 
-// Store instance
+// Store singleton
 let store: LazyStore | null = null;
 
 // Svelte store for real-time updates
@@ -18,16 +18,13 @@ export const clipboardHistory: Writable<ClipboardHistoryItem[]> = writable([]);
  * Initialize the clipboard history store
  */
 export async function initClipboardStore(): Promise<void> {
-  if (!store) {
-    store = new LazyStore(STORE_PATH, {
-      autoSave: 100, // Auto save with 100ms debounce
-    });
-    await store.init();
+  if (store) return;
 
-    // Load initial data
-    const items = await getHistoryItems();
-    clipboardHistory.set(items);
-  }
+  store = new LazyStore(STORE_PATH, { autoSave: 100 });
+  await store.init();
+
+  // Load initial data
+  clipboardHistory.set(await getHistoryItems());
 }
 
 /**
@@ -36,15 +33,12 @@ export async function initClipboardStore(): Promise<void> {
 export async function addHistoryItem(
   item: ClipboardHistoryItem
 ): Promise<void> {
-  if (!store) {
-    await initClipboardStore();
-  }
+  if (!store) await initClipboardStore();
 
   try {
-    // Get current items
     let items = await getHistoryItems();
 
-    // Check for duplicates (by content for text/html, by id for images)
+    // Check for duplicates by content or ID
     const isDuplicate = items.some(
       (existing) =>
         item.type === existing.type &&
@@ -52,22 +46,18 @@ export async function addHistoryItem(
           (item.type === "image" && item.id === existing.id))
     );
 
-    if (!isDuplicate) {
-      // Add new item at the beginning
-      items = [item, ...items];
+    if (isDuplicate) return;
 
-      // Clean up old items
-      const cutoffTime = Date.now() - MAX_HISTORY_AGE_MS;
-      items = items
-        .filter((item) => item.createdAt > cutoffTime || item.favorite)
-        .slice(0, MAX_ITEMS);
+    // Add new item and clean up old ones
+    items = [item, ...items];
 
-      // Save to store
-      await store?.set("items", items);
+    const cutoffTime = Date.now() - MAX_HISTORY_AGE_MS;
+    items = items
+      .filter((item) => item.favorite || item.createdAt > cutoffTime)
+      .slice(0, MAX_ITEMS);
 
-      // Update Svelte store
-      clipboardHistory.set(items);
-    }
+    await store?.set("items", items);
+    clipboardHistory.set(items);
   } catch (error) {
     LogService.error(`Failed to add clipboard history item: ${error}`);
   }
@@ -77,13 +67,10 @@ export async function addHistoryItem(
  * Get all clipboard history items
  */
 export async function getHistoryItems(): Promise<ClipboardHistoryItem[]> {
-  if (!store) {
-    await initClipboardStore();
-  }
+  if (!store) await initClipboardStore();
 
   try {
-    const items = (await store?.get<ClipboardHistoryItem[]>("items")) || [];
-    return items;
+    return (await store?.get<ClipboardHistoryItem[]>("items")) || [];
   } catch (error) {
     LogService.error(`Failed to get clipboard history items: ${error}`);
     return [];
@@ -94,9 +81,7 @@ export async function getHistoryItems(): Promise<ClipboardHistoryItem[]> {
  * Toggle favorite status of an item
  */
 export async function toggleFavorite(id: string): Promise<void> {
-  if (!store) {
-    await initClipboardStore();
-  }
+  if (!store) await initClipboardStore();
 
   try {
     const items = await getHistoryItems();
@@ -115,9 +100,7 @@ export async function toggleFavorite(id: string): Promise<void> {
  * Delete an item from history
  */
 export async function deleteHistoryItem(id: string): Promise<void> {
-  if (!store) {
-    await initClipboardStore();
-  }
+  if (!store) await initClipboardStore();
 
   try {
     const items = await getHistoryItems();
@@ -134,9 +117,7 @@ export async function deleteHistoryItem(id: string): Promise<void> {
  * Clear all non-favorite items from history
  */
 export async function clearHistory(): Promise<void> {
-  if (!store) {
-    await initClipboardStore();
-  }
+  if (!store) await initClipboardStore();
 
   try {
     const items = await getHistoryItems();
