+++ src/types/index.ts
@@ -1,31 +1,14 @@
-import type { ActionResult } from "./actionResult";
-import type { AppResult } from "./appResult";
-import type { CategoryResults } from "./categoryResults";
-import type { SearchResultItem } from "./searchResultItem";
-import type { SearchResults } from "./searchResults";
-import type { ResultCategory } from "./resultCategory";
-import type {
+export type { ActionResult } from "./actionResult";
+export type { AppResult } from "./appResult";
+export type { CategoryResults } from "./categoryResults";
+export type { SearchResultItem } from "./searchResultItem";
+export type { SearchResults } from "./searchResults";
+export type { ResultCategory } from "./resultCategory";
+export type {
   Extension,
   ExtensionManifest,
   ExtensionCommand,
   ExtensionResult,
 } from "./extension";
-import type {
-  ClipboardHistoryItem,
-  ClipboardItemType,
-} from "./clipboardHistoryItem";
-
-export type {
-  ActionResult,
-  AppResult,
-  CategoryResults,
-  SearchResultItem,
-  SearchResults,
-  ResultCategory,
-  Extension,
-  ExtensionManifest,
-  ExtensionCommand,
-  ExtensionResult,
-  ClipboardHistoryItem,
-  ClipboardItemType,
-};
+export type { ClipboardHistoryItem } from "./clipboardHistoryItem";
+export { ClipboardItemType } from "./clipboardHistoryItem";
